@@ -65,7 +65,7 @@
     "@custom-elements-manifest/analyzer": "0.11.0",
     "@custom-elements-manifest/to-markdown": "0.1.0",
     "@eslint/eslintrc": "3.3.3",
-    "@eslint/js": "9.39.2",
+    "@eslint/js": "10.0.1",
     "@inquirer/checkbox": "5.0.4",
     "@lit-labs/observers": "2.1.0",
     "@lit-labs/router": "0.1.4",
@@ -93,7 +93,7 @@
     "custom-elements-manifest": "2.1.0",
     "date-fns": "4.1.0",
     "esbuild": "0.27.3",
-    "eslint": "9.39.2",
+    "eslint": "10.0.0",
     "eslint-config-prettier": "10.1.8",
     "eslint-import-resolver-typescript": "4.4.4",
     "eslint-plugin-import-x": "4.16.1",
